backend/functions/stripeWebHookHandler/requirements.txt
firebase-admin==6.1.0
google-cloud-secret-manager==2.16.1
Jinja2==3.0.3
python-dateutil==2.8.1
stripe==5.4.0
Werkzeug